Move Paw class to core.py and upgrade to 2.0.0
* Move Paw class implementation from __init__.py to its own module
* Upgrade to version 2.0.0 to align with wlgen 2.0.0 dependency
* No functional changes or breaking API changes

Part of modernization effort to improve code organization.
