+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name="paw"
-version="1.3.3"
+version="2.0.0"
 description="paw - patterns and wordlists in python"
 readme = "README.md"
 authors = [
@@ -8,7 +8,7 @@ authors = [
 ]
 requires-python = ">=3.0"
 dependencies = [
-    "wlgen>=1.3.3",
+    "wlgen>=2.0.0",
 ]
 [dependency-groups]
 dev = [
